Fix build

+++ src/serialize/ParserProcessValuesInOrder.cpp
@@ -193,7 +193,7 @@ void ParserProcessValuesInOrder::Internal::exitArrayStruct(const MetaField& fiel
             m_currentCalls = &m_stackCalls.back();
             assert(!m_currentCalls->arr);
             assert(field.index >= 0);
-            if ((field.index + 1) < m_currentCalls->calls.size())
+            if ((field.index + 1) < static_cast<int>(m_currentCalls->calls.size()))
             {
                 m_currentCalls->calls[field.index + 1] = callsSub;
             }
